tsl2580: Add a simple driver for tsl2580
Driver consists of the source file, dts overlay and registers description.
When the device is detected, it's initialized with a default configuration,
that can be seen in the source file. It also has a single sysfs entry,
which reads from the device its id, and prints what type of device is
attached, if it's supported.

Signed-off-by: Andrii Revva <milaner.work@gmail.com>