tsl2580: Add a lux attribute in the sysfs class
It seems like a redundant thing, because to precicely calculate lux values
from adc0 and adc1 it's beneficial to have floating point arithmetic. So
it would be nice to pass this obligation to user space.
The method to approximately calculate lux values comes from the datasheet
and it is _absolutely_ arcane and obscure.

Signed-off-by: Andrii Revva <milaner.work@gmail.com>